:pick: Cast `int` result to `uint8_t` for PID calc
`-Wconversion` FTW

+++ src/lin_pid.c
@@ -93,8 +93,8 @@ uint8_t ComputePID(uint8_t id)
    // 
    // So, the eventual bits will look like, MSb first:
    //   P1 P0 ID5 ID4 ID3 ID2 ID1 ID0
-   pid |= ( ( GET_BIT(pid, 0) ^ GET_BIT(pid, 1) ^ GET_BIT(pid, 2) ^ GET_BIT(pid, 4) ) << 6 );
-   pid |= (!( GET_BIT(pid, 1) ^ GET_BIT(pid, 3) ^ GET_BIT(pid, 4) ^ GET_BIT(pid, 5) ) << 7 );
+   pid |= (uint8_t)( ( GET_BIT(pid, 0) ^ GET_BIT(pid, 1) ^ GET_BIT(pid, 2) ^ GET_BIT(pid, 4) ) << 6 );
+   pid |= (uint8_t)(!( GET_BIT(pid, 1) ^ GET_BIT(pid, 3) ^ GET_BIT(pid, 4) ^ GET_BIT(pid, 5) ) << 7 );
 
    return pid;
 }
